Restructuring and renaming + minor fixes
